fixup! feat(leaderelection): add lease/TTL strategy behind a selector flag
